requirements.txt
attrs==25.3.0
autobahn==19.11.2
Automat==25.4.16
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.5
curl_cffi==0.12.0
cycler==0.12.1
db-dtypes==1.4.3
fonttools==4.59.0
frozendict==2.4.6
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
joblib==1.5.1
kiteconnect==5.0.1
kiwisolver==1.4.8
matplotlib==3.10.3
multitasking==0.0.11
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
pandas-gbq==0.29.2
pandas_ta==0.3.14b0
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydata-google-auth==1.9.1
Pygments==2.19.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
service-identity==24.2.0
setuptools==80.9.0
# Editable Git install with no remote (shared-services==0.1.0)
-e /Users/sivarajumalladi/Documents/GitHub/shared-services
six==1.17.0
soupsieve==2.7
# Editable Git install with no remote (strategy-engine==0.1.0)
-e /Users/sivarajumalladi/Documents/GitHub/strategy-engine
ta-lib==0.6.4
threadpoolctl==3.6.0
tqdm==4.67.1
# Editable Git install with no remote (trading-data-pipeline==0.1.0)
-e /Users/sivarajumalladi/Documents/GitHub/trading-data-pipeline
# Editable Git install with no remote (trading-execution-engine==0.1.0)
-e /Users/sivarajumalladi/Documents/GitHub/trading-execution-engine
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.65
zope.interface==7.2
